fix(web): surface server rejection message in contribute error toast
Contribute onError was swallowing the server message and showing a
generic "Failed to contribute photo", hiding actionable feedback like
"This photo has already been contributed". Use getMutationErrorMessage
to pull the real message from the ApiError body.

Promotes getMutationErrorMessage/isBannerError from admin/lib to
@/lib/api-errors since nothing about them is admin-specific; the
admin re-export keeps existing imports working.
