+++ web/src/lib/api-errors.ts
@@ -1,0 +1,25 @@
+import { ApiError } from '@/lib/api-client';
+
+/**
+ * Returns true when the error is an ApiError with a status code that should
+ * be displayed inline (banner or toast) rather than swallowed as a generic
+ * "something went wrong".
+ *
+ * Used to distinguish "the server is telling you something actionable"
+ * (400/403/404/409) from "something failed transiently" (5xx, network).
+ */
+export function isBannerError(err: unknown): err is ApiError {
+  return err instanceof ApiError && [400, 403, 404, 409].includes(err.status);
+}
+
+/**
+ * Extracts a human-readable message from an error. For banner-class
+ * ApiErrors, returns the server-provided message; otherwise returns a
+ * generic fallback.
+ */
+export function getMutationErrorMessage(err: unknown): string {
+  if (isBannerError(err)) {
+    return err.body.error;
+  }
+  return 'An unexpected error occurred. Please try again.';
+}
